% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{Accounts}
\alias{Accounts}
\title{Accounts Object}
\usage{
Accounts(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The accounts returned in this list response}

\item{nextPageToken}{Continuation token used to page through accounts}
}
\value{
Accounts object
}
\description{
Accounts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
