% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{AdClient}
\alias{AdClient}
\title{AdClient Object}
\usage{
AdClient(arcOptIn = NULL, id = NULL, productCode = NULL,
  supportsReporting = NULL)
}
\arguments{
\item{arcOptIn}{Whether this ad client is opted in to ARC}

\item{id}{Unique identifier of this ad client}

\item{productCode}{This ad client's product code, which corresponds to the PRODUCT_CODE report dimension}

\item{supportsReporting}{Whether this ad client supports being reported on}
}
\value{
AdClient object
}
\description{
AdClient Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

