% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{PreferredDeal}
\alias{PreferredDeal}
\title{PreferredDeal Object}
\usage{
PreferredDeal(advertiserName = NULL, buyerNetworkName = NULL,
  currencyCode = NULL, endTime = NULL, fixedCpm = NULL, id = NULL,
  startTime = NULL)
}
\arguments{
\item{advertiserName}{The name of the advertiser this deal is for}

\item{buyerNetworkName}{The name of the buyer network this deal is for}

\item{currencyCode}{The currency code that applies to the fixed_cpm value}

\item{endTime}{Time when this deal stops being active in seconds since the epoch (GMT)}

\item{fixedCpm}{The fixed price for this preferred deal}

\item{id}{Unique identifier of this preferred deal}

\item{startTime}{Time when this deal becomes active in seconds since the epoch (GMT)}
}
\value{
PreferredDeal object
}
\description{
PreferredDeal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
