% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{UrlChannels}
\alias{UrlChannels}
\title{UrlChannels Object}
\usage{
UrlChannels(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The URL channels returned in this list response}

\item{nextPageToken}{Continuation token used to page through URL channels}
}
\value{
UrlChannels object
}
\description{
UrlChannels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

